+++ include/geos/algorithm/RayCrossingCounter.h
@@ -65,7 +65,7 @@ class GEOS_DLL RayCrossingCounter {
 private:
     const geom::Coordinate& point;
 
-    int crossingCount;
+    std::size_t crossingCount;
 
     // true if the test point lies on an input segment
     bool isPointOnSegment;
@@ -143,6 +143,8 @@ public:
      * @return `true` if the point lies in or on the supplied polygon
      */
     bool isPointInPolygon() const;
+
+    std::size_t getCount() const { return crossingCount; };
 
 };