+++ include/geos/simplify/TaggedLineString.h
@@ -47,6 +47,9 @@ class TaggedLineSegment;
 }
 }
 
+using geos::geom::Coordinate;
+using geos::geom::CoordinateSequence;
+
 namespace geos {
 namespace simplify { // geos::simplify
 
@@ -58,26 +61,35 @@ class GEOS_DLL TaggedLineString {
 
 public:
 
-    typedef std::vector<geom::Coordinate> CoordVect;
+    typedef std::vector<Coordinate> CoordVect;
 
     typedef std::unique_ptr<CoordVect> CoordVectPtr;
 
-    typedef geom::CoordinateSequence CoordSeq;
+    typedef CoordinateSequence CoordSeq;
 
-    typedef std::unique_ptr<geom::CoordinateSequence> CoordSeqPtr;
+    typedef std::unique_ptr<CoordinateSequence> CoordSeqPtr;
 
     TaggedLineString(const geom::LineString* nParentLine,
-                     std::size_t minimumSize = 2);
+                     std::size_t minimumSize,
+                     bool bIsRing);
 
     ~TaggedLineString();
 
     std::size_t getMinimumSize() const;
+
+    bool isRing() const;
 
     const geom::LineString* getParent() const;
 
     const CoordSeq* getParentCoordinates() const;
 
     CoordSeqPtr getResultCoordinates() const;
+
+    const Coordinate& getCoordinate(std::size_t i) const;
+
+    std::size_t size() const;
+
+    const Coordinate& getComponentPoint() const;
 
     std::size_t getResultSize() const;
 
@@ -89,7 +101,11 @@ public:
 
     const std::vector<TaggedLineSegment*>& getSegments() const;
 
+    const std::vector<TaggedLineSegment*>& getResultSegments() const;
+
     void addToResult(std::unique_ptr<TaggedLineSegment> seg);
+
+    const TaggedLineSegment* removeRingEndpoint();
 
     std::unique_ptr<geom::Geometry> asLineString() const;
 
@@ -107,9 +123,11 @@ private:
 
     std::size_t minimumSize;
 
+    bool m_isRing;
+
     void init();
 
-    static CoordVectPtr extractCoordinates(
+    static std::unique_ptr<CoordinateSequence> extractCoordinates(
         const std::vector<TaggedLineSegment*>& segs);
 
     // Copying is turned off